+++ src/Model/File.hs
@@ -1,0 +1,12 @@
+module Model.File
+  ( module Model.FileBookmark,
+    module Model.FileFirefox,
+    module Model.FileNetscape,
+    module Model.FileNote,
+  )
+where
+
+import Model.FileBookmark
+import Model.FileFirefox
+import Model.FileNetscape
+import Model.FileNote
